renderer code moved out of main

+++ pkg/indexBuffer/indexBuffer.go
@@ -9,7 +9,7 @@ import (
 
 type IndexBuffer struct {
 	RendererID uint32
-	Count int
+	Count int32
 }
 
 func New(data interface{}, count int) *IndexBuffer {
@@ -18,7 +18,7 @@ func New(data interface{}, count int) *IndexBuffer {
 	gl.GenBuffers(1, &ib.RendererID);
 	gl.BindBuffer(gl.ELEMENT_ARRAY_BUFFER, ib.RendererID)
 	gl.BufferData(gl.ELEMENT_ARRAY_BUFFER, int(count * 32), gl.Ptr(data), gl.STATIC_DRAW)
-	ib.Count = count
+	ib.Count = int32(count)
 	log.Println(fmt.Sprintf("Index Buffer %d added", ib.RendererID))
 
 	return &ib
@@ -36,6 +36,6 @@ func (ib *IndexBuffer) Unbind() {
 	gl.BindBuffer(gl.ELEMENT_ARRAY_BUFFER, 0)
 }
 
-func (ib *IndexBuffer) GetCount() int {
+func (ib *IndexBuffer) GetCount() int32 {
 	return ib.Count
 }
